requirements.txt
aiohttp
daphne
Django
django-heroku
python-decouple

